backend/requirements-dev.txt
# Development dependencies for NoobBook Backend
# Testing
pytest==9.0.1
pytest-asyncio==0.26.2
pytest-cov==6.0.0
pytest-mock==3.16.0
pytest-flask==1.3.0

# Code Quality
black==25.11.0
flake8==7.3.0
isort==6.1.3
mypy==1.18.0

# Type stubs
types-requests==2.32.0.20250117
types-python-dateutil==2.9.0.20250117
